Initial Upload
matrix is a class for an array representing a two dimensional matrix with indexes starting at 1. The array size must be provided at object construction. The row and column indices are assumed from the declared matrix size and the array index of the matrix element.

Ex. an element in the 2,1 index in a 3x3 matrix is the 4th element in the containing array.